requirements.txt
absl-py==2.3.0
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
datasets==3.2.0
dill==0.3.8
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.73.0
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.5
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
mlx==0.22.1
mlx-lm==0.21.4
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
peft==0.14.0
portalocker==3.2.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sacrebleu==2.5.1
safetensors==0.5.2
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.23.0
